moved js files to modules folder
