+++ public/js/modules/stocksTheme.js
